tools/requirements.txt
dandi
pyyaml
selenium

# Non-Python requirement: chromedriver
